Use Celery task for create_subscription route
This also moves the logic for building a subscription payload to the
jobs module
