+++ atat/jobs.py
@@ -1,3 +1,4 @@
+from secrets import token_urlsafe
 from smtplib import SMTPException
 
 import pendulum
@@ -8,17 +9,18 @@ from flask import current_app as app
 from atat.database import db
 from atat.domain.application_roles import ApplicationRoles
 from atat.domain.applications import Applications
-from atat.domain.environment_roles import EnvironmentRoles
 from atat.domain.csp.cloud import CloudProviderInterface
 from atat.domain.csp.cloud.exceptions import GeneralCSPException
 from atat.domain.csp.cloud.models import (
     ApplicationCSPPayload,
     BillingInstructionCSPPayload,
     EnvironmentCSPPayload,
+    SubscriptionCreationCSPPayload,
     UserCSPPayload,
     UserRoleCSPPayload,
 )
 from atat.domain.csp.cloud.utils import generate_user_principal_name
+from atat.domain.environment_roles import EnvironmentRoles
 from atat.domain.environments import Environments
 from atat.domain.portfolios import Portfolios
 from atat.domain.task_orders import TaskOrders
@@ -194,6 +196,27 @@ def create_subscription(self, environment_id=None):
     return environment_id
 
 
+def build_subscription_payload(environment) -> SubscriptionCreationCSPPayload:
+    csp_data = environment.portfolio.csp_data
+    parent_group_id = environment.cloud_id
+    invoice_section_name = csp_data["billing_profile_properties"]["invoice_sections"][
+        0
+    ]["invoice_section_name"]
+
+    display_name = (
+        f"{environment.application.name}-{environment.name}-{token_urlsafe(6)}"
+    )
+
+    return SubscriptionCreationCSPPayload(
+        tenant_id=csp_data.get("tenant_id"),
+        display_name=display_name,
+        parent_group_id=parent_group_id,
+        billing_account_name=app.config["AZURE_BILLING_ACCOUNT_NAME"],
+        billing_profile_name=csp_data.get("billing_profile_name"),
+        invoice_section_name=invoice_section_name,
+    )
+
+
 def do_create_subscription(csp: CloudProviderInterface, environment_id=None):
     """Creates a subscription under a management group for an environment
     
@@ -202,9 +225,7 @@ def do_create_subscription(csp: CloudProviderInterface, environment_id=None):
     that a request to kick off this async job is accepted.
     """
     environment = Environments.get(environment_id)
-    payload = environment.build_subscription_payload(
-        app.config["AZURE_BILLING_ACCOUNT_NAME"]
-    )
+    payload = build_subscription_payload(environment)
     try:
         csp.create_subscription(payload)
     except GeneralCSPException as e:
